(Chore) Rename media
Changed Media to ts
